libpanel::subclass

Module prelude

Re-exports

pub use super::application::PanelApplicationImpl;

pub use super::dock::DockImpl;

pub use super::dock::DockImplExt;

pub use super::document_workspace::DocumentWorkspaceImpl;

pub use super::document_workspace::DocumentWorkspaceImplExt;

pub use super::frame::PanelFrameImpl;

pub use super::frame::PanelFrameImplExt;

pub use super::frame_header::FrameHeaderImpl;

pub use super::frame_header::FrameHeaderImplExt;

pub use super::grid::PanelGridImpl;

pub use super::grid::PanelGridImplExt;

pub use super::omni_bar::OmniBarImpl;

pub use super::save_delegate::SaveDelegateImpl;

pub use super::save_delegate::SaveDelegateImplExt;

pub use super::widget::PanelWidgetClassSubclassExt;

pub use super::widget::PanelWidgetImpl;

pub use super::widget::PanelWidgetImplExt;

pub use super::workbench::WorkbenchClassSubclassExt;

pub use super::workbench::WorkbenchImpl;

pub use super::workbench::WorkbenchImplExt;

pub use super::workspace::WorkspaceClassSubclassExt;

pub use super::workspace::WorkspaceImpl;

Structs

TemplateChild

Traits

AccessibleImpl

AccessibleImplExt

AccessibleRangeImpl

AccessibleRangeImplExt

AccessibleTextImpl

AccessibleTextImplExt

ActionGroupImpl

ActionGroupImplExt

ActionMapImpl

ActionMapImplExt

ActionableImpl

ActionableImplExt

AdjustmentImpl

AdjustmentImplExt

ApplicationImpl

ApplicationImplExt

ApplicationWindowImpl

AsyncInitableImpl

AsyncInitableImplExt

BoxImpl

BoxedType

Trait for defining boxed types.

BuildableImpl

BuildableImplExt

BuilderCScopeImpl

BuilderScopeImpl

BuilderScopeImplExt

ButtonImpl

ButtonImplExt

CellAreaClassExt

CellAreaContextImplDeprecated

CellAreaContextImplExtDeprecated

CellAreaImplDeprecated

CellAreaImplExtDeprecated

CellEditableImplDeprecated

CellEditableImplExtDeprecated

CellLayoutImplDeprecated

CellLayoutImplExtDeprecated

CellRendererImplDeprecated

CellRendererImplExtDeprecated

CellRendererTextImplDeprecated

CellRendererTextImplExtDeprecated

CheckButtonImpl

CheckButtonImplExt

ClassStruct

Trait implemented by structs that implement a GObject C class struct.

ColorChooserImplDeprecated

ColorChooserImplExtDeprecated

ComboBoxImplDeprecated

ComboBoxImplExtDeprecated

CompositeTemplate

A trait for setting up template children inside class_init. This trait is implemented automatically by the CompositeTemplate macro.

CompositeTemplateCallbacks

A trait for setting up template callbacks inside class_init. This trait is implemented automatically by the template_callbacks macro.

CompositeTemplateCallbacksClass

An extension trait for ClassStruct types to allow binding private template callbacks directly on self. This is a convenience wrapper around the CompositeTemplateCallbacks trait.

CompositeTemplateClass

An extension trait for ClassStruct types to allow binding a composite template directly on self. This is a convenience wrapper around the CompositeTemplate trait.

CompositeTemplateDisposeExt

CompositeTemplateInitializingExt

CompositeTemplateInstanceCallbacksClass

An extension trait for ClassStruct types to allow binding the instance template callbacks directly on self. This is a convenience wrapper around the CompositeTemplateCallbacks trait.

ContentProviderImpl

ContentProviderImplExt

DerivedObjectProperties

Trait containing only the property related functions of ObjectImpl. Implemented by the Properties macro. When implementing ObjectImpl you may want to delegate the function calls to this trait.

DialogImpl

DialogImplExt

DrawingAreaImpl

DrawingAreaImplExt

EditableImpl

EditableImplExt

EntryBufferImpl

EntryBufferImplExt

EntryImpl

EntryImplExt

FileEnumeratorImpl

FileEnumeratorImplExt

FileMonitorImpl

FileMonitorImplExt

FilterImpl

FilterImplExt

FixedImpl

FlowBoxChildImpl

FlowBoxChildImplExt

FontChooserImplDeprecated

FontChooserImplExtDeprecated

FrameImplDeprecated

FrameImplExtDeprecated

GLAreaImpl

GLAreaImplExt

GridImpl

GtkApplicationImpl

GtkApplicationImplExt

IMContextImpl

IMContextImplExt

IOStreamImpl

IOStreamImplExt

InitableImpl

InitableImplExt

InputStreamImpl

InputStreamImplExt

InstanceStruct

Trait implemented by structs that implement a GObject C instance struct.

InstanceStructExt

Trait implemented by any type implementing InstanceStruct to return the implementation, private Rust struct.

InterfaceStruct

Trait implemented by structs that implement a GTypeInterface C class struct.

IsImplementable

Trait for implementable interfaces.

IsSubclassable

Trait for subclassable class structs.

IsSubclassableExt

LayoutChildImpl

LayoutManagerImpl

LayoutManagerImplExt

ListBoxRowImpl

ListBoxRowImplExt

ListModelImpl

ListModelImplExt

MediaFileImpl

MediaFileImplExt

MediaStreamImpl

MediaStreamImplExt

NativeDialogImplDeprecated

NativeDialogImplExtDeprecated

ObjectClassSubclassExt

Extension trait for glib::Object’s class struct.

ObjectImpl

Trait for implementors of glib::Object subclasses.

ObjectImplExt

ObjectInterface

The central trait for defining a GObject interface.

ObjectInterfaceExt

ObjectInterfaceType

Type methods required for an ObjectInterface implementation.

ObjectSubclass

The central trait for subclassing a GObject type.

ObjectSubclassExt

Extension methods for all ObjectSubclass impls.

ObjectSubclassIsExt

Trait implemented by any type implementing ObjectSubclassIs to return the implementation, private Rust struct.

ObjectSubclassType

Type methods required for an ObjectSubclass implementation.

OrientableImpl

OutputStreamImpl

OutputStreamImplExt

PaintableImpl

PaintableImplExt

PopoverImpl

PopoverImplExt

PrintOperationImpl

PrintOperationImplExt

PrintOperationPreviewImpl

RangeImpl

RangeImplExt

RecentManagerImpl

RecentManagerImplExt

RefCounted

ScaleButtonImpl

ScaleButtonImplExt

ScaleImpl

ScaleImplExt

ScrollableImpl

ScrollableImplExt

SectionModelImpl

SectionModelImplExt

SeekableImpl

SeekableImplExt

SelectionModelImpl

SelectionModelImplExt

SharedType

Trait for defining shared types.

ShortcutManagerImpl

ShortcutManagerImplExt

SocketControlMessageImpl

SocketControlMessageImplExt

SorterImpl

SorterImplExt

StyleContextImplDeprecated

StyleContextImplExtDeprecated

SymbolicPaintableImpl

SymbolicPaintableImplExt

TextBufferImpl

TextBufferImplExt

TextViewImpl

TextViewImplExt

ToggleButtonImpl

ToggleButtonImplExt

TreeDragDestImplDeprecated

TreeDragDestImplExtDeprecated

TreeDragSourceImplDeprecated

TreeDragSourceImplExtDeprecated

TreeModelFilterImplDeprecated

TreeModelFilterImplExtDeprecated

TreeViewImplDeprecated

TreeViewImplExtDeprecated

TypeModuleImpl

TypeModuleImplExt

TypePluginImpl

TypePluginImplExt

TypePluginRegisterImpl

VfsImpl

VfsImplExt

WidgetClassExt

WidgetImpl

WidgetImplExt

WindowGroupImpl

WindowImpl

WindowImplExt