Module libpanel::subclass::prelude

Re-exports

pub use super::dock::DockImpl;

pub use super::dock::DockImplExt;

pub use super::frame::PanelFrameImpl;

pub use super::frame_header::FrameHeaderImpl;

pub use super::frame_header::FrameHeaderImplExt;

pub use super::grid::PanelGridImpl;

pub use super::grid::PanelGridImplExt;

pub use super::omni_bar::PanelOmniBarImpl;

pub use super::save_delegate::SaveDelegateImpl;

pub use super::save_delegate::SaveDelegateImplExt;

pub use super::widget::PanelWidgetImpl;

pub use super::widget::PanelWidgetImplExt;

Structs

TemplateChild

Traits

AccessibleRangeImpl

AccessibleRangeImplExt

ActionGroupImpl

ActionGroupImplExt

ActionMapImpl

ActionMapImplExt

ActionableImpl

ActionableImplExt

AdjustmentImpl

AdjustmentImplExt

ApplicationImpl

ApplicationImplExt

ApplicationWindowImpl

AsyncInitableImpl

AsyncInitableImplExt

BoxImpl

BoxedType

Trait for defining boxed types.

BuildableImpl

BuildableImplExt

BuilderCScopeImpl

BuilderScopeImpl

BuilderScopeImplExt

ButtonImpl

ButtonImplExt

CellAreaClassSubclassExt

CellAreaContextImpl

CellAreaContextImplExt

CellAreaImpl

CellAreaImplExt

CellEditableImpl

CellEditableImplExt

CellLayoutImpl

CellLayoutImplExt

CellRendererImpl

CellRendererImplExt

CellRendererTextImpl

CellRendererTextImplExt

CheckButtonImpl

CheckButtonImplExt

ClassStruct

Trait implemented by structs that implement a GObject C class struct.

ColorChooserImpl

ColorChooserImplExt

ComboBoxImpl

ComboBoxImplExt

CompositeTemplate

A trait for setting up template children inside class_init. This trait is implemented automatically by the CompositeTemplate macro.

CompositeTemplateCallbacks

A trait for setting up template callbacks inside class_init. This trait is implemented automatically by the template_callbacks macro.

CompositeTemplateCallbacksClass

An extension trait for ClassStruct types to allow binding private template callbacks directly on self. This is a convenience wrapper around the CompositeTemplateCallbacks trait.

CompositeTemplateClass

An extension trait for ClassStruct types to allow binding a composite template directly on self. This is a convenience wrapper around the CompositeTemplate trait.

CompositeTemplateInstanceCallbacksClass

An extension trait for ClassStruct types to allow binding the instance template callbacks directly on self. This is a convenience wrapper around the CompositeTemplateCallbacks trait.

ConstraintTargetImpl

ContentProviderImpl

ContentProviderImplExt

DialogImpl

DialogImplExt

DrawingAreaImpl

DrawingAreaImplExt

EditableImpl

EditableImplExt

EntryBufferImpl

EntryBufferImplExt

EntryImpl

EntryImplExt

FilterImpl

FilterImplExt

FixedImpl

FlowBoxChildImpl

FlowBoxChildImplExt

FontChooserImpl

FontChooserImplExt

FrameImpl

FrameImplExt

GLAreaImpl

GLAreaImplExt

GridImpl

GtkApplicationImpl

GtkApplicationImplExt

IMContextImpl

IMContextImplExt

IOStreamImpl

IOStreamImplExt

InitableImpl

InitableImplExt

InputStreamImpl

InputStreamImplExt

InstanceStruct

Trait implemented by structs that implement a GObject C instance struct.

InstanceStructExt

Trait implemented by any type implementing InstanceStruct to return the implementation, private Rust struct.

IsImplementable

Trait for implementable interfaces.

IsSubclassable

Trait for subclassable class structs.

IsSubclassableExt

LayoutChildImpl

LayoutManagerImpl

LayoutManagerImplExt

ListBoxRowImpl

ListBoxRowImplExt

ListModelImpl

ListModelImplExt

MediaFileImpl

MediaFileImplExt

MediaStreamImpl

MediaStreamImplExt

NativeDialogImpl

NativeDialogImplExt

NativeImpl

ObjectClassSubclassExt

Extension trait for glib::Object’s class struct.

ObjectImpl

Trait for implementors of glib::Object subclasses.

ObjectImplExt

ObjectInterface

The central trait for defining a GObject interface.

ObjectInterfaceExt

ObjectInterfaceType

Type methods required for an ObjectInterface implementation.

ObjectSubclass

The central trait for subclassing a GObject type.

ObjectSubclassExt

Extension methods for all ObjectSubclass impls.

ObjectSubclassIsExt

Trait implemented by any type implementing ObjectSubclassIs to return the implementation, private Rust struct.

ObjectSubclassType

Type methods required for an ObjectSubclass implementation.

OrientableImpl

OutputStreamImpl

OutputStreamImplExt

PaintableImpl

PaintableImplExt

PopoverImpl

PopoverImplExt

PrintOperationImpl

PrintOperationImplExt

PrintOperationPreviewImpl

RangeImpl

RangeImplExt

RecentManagerImpl

RecentManagerImplExt

RefCounted

RootImpl

ScaleButtonImpl

ScaleButtonImplExt

ScaleImpl

ScaleImplExt

ScrollableImpl

ScrollableImplExt

SeekableImpl

SeekableImplExt

SelectionModelImpl

SelectionModelImplExt

SharedType

Trait for defining shared types.

ShortcutManagerImpl

ShortcutManagerImplExt

SorterImpl

SorterImplExt

StyleContextImpl

StyleContextImplExt

SymbolicPaintableImpl

SymbolicPaintableImplExt

TextBufferImpl

TextBufferImplExt

TextViewImpl

TextViewImplExt

ToggleButtonImpl

ToggleButtonImplExt

TreeDragDestImpl

TreeDragDestImplExt

TreeDragSourceImpl

TreeDragSourceImplExt

TreeModelFilterImpl

TreeModelFilterImplExt

TreeViewImpl

TreeViewImplExt

WidgetClassSubclassExt

WidgetImpl

WidgetImplExt

WindowImpl

WindowImplExt